+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - __Fixes:__
   - Slider velocity is now adjusted properly for taiko converts
+  - Fixed missing slider sounds for taiko converts
 - __Breaking changes:__
   - Replaced the simple `Strains` struct with a new struct `{Mode}Strains` that contains more detail w.r.t. the mode.
   - Renamed all `GameMode` variants to more idiomatic names